feat: refactor account structures to use MerkleContextV2 and update API schemas

+++ src/ingester/parser/tx_event_parser.rs
@@ -164,7 +164,7 @@ pub fn parse_public_transaction_event(
             false,
             None,
             None,
-            tree_and_queue.tree_type as u16,
+            tree_and_queue.tree_type,
         );
 
         state_update.out_accounts.push(enriched_account);
